#1985: refactor: Moves SaveButtonWithDropdown to separate directory
