+++ src/components/SlateEditor/EditorFooter.tsx
@@ -14,12 +14,11 @@ import { colors, spacing } from '@ndla/core';
 import { Launch } from '@ndla/icons/common';
 
 import { toPreviewDraft } from '../../util/routeHelpers';
-import SaveButton from '../SaveButton';
 import { Article, PossibleStatuses, Values } from './editorTypes';
 import * as draftApi from '../../modules/draft/draftApi';
 import { formatErrorMessage } from '../../util/apiHelpers';
 import { TranslateType } from '../../interfaces';
-import SaveButtonDropDown from '../SaveButtonDropDown';
+import SaveButtonDropDown from '../SaveButtonWithDropDown/SaveButtonWithDropdown';
 
 interface Props {
   t: TranslateType;
